+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Improved Blizzard UI Clean Changelog
+# Improved Blizzard UI Clean Edition Changelog
+1.1.4 - 
+* Framework added for Action Bar Art (Gryphon end-caps). It is now possible to remove the end caps from the action bar, more functions to come later!
+* ToT (Target of Target) reduced font size to improve readability of large nameplates.
 1.1.3 - Minimap Zone Text better fits the larger zone names.
 1.1.2 - Chat now uses uppercase characters for representing Guild Party, e.t.c. The text box font size is slightly increased.
 1.1.1 - Re-applied adjustments from prev versions. (Brute Force Frame Pos??, I think not.)
@@ -6,3 +9,4 @@
 1.0.5 - Framework for Gryphon Art Hide added to config menu, Proper Frame Unlock and removed hacky fix.
 1.0.1 - Tweaks to frame & misc font default sizes for better fit, hacky unit frame unlock fix.
 1.0.0 - Unlinked interface font from chat. Re-posionted focus frame. 
+Project Start (BFA 8.0 Base)